+++ CHANGES.md
@@ -5,6 +5,8 @@ E-Maj - Change log
 ###Enhancements:###
   * E-Maj is compatible with PostgreSQL V18, including the support of virtual
     generated columns.
+  * Add 2 functions sets to show changes statistics for a single table or
+    sequence on a given time frame.
   * Add checks on the stability of "GENERATED ALWAYS AS expression" columns
     in application tables.
   * Add a check on the existence of application triggers registered as "to be